workflows/requirements.txt
rocrate~=0.4.0
planemo>=0.74.5
pyyaml~=5.4.0
requests~=2.26.0
